feat(main-page): main page categories
