--- client/src/Content.css
@@ -1,12 +1,0 @@
-.Content
-{
-    display: flex;
-    min-height: 100vh;
-}
-.Content-product
-{
-    display: grid;
-    grid-template-rows: 1fr 1fr 1fr 1fr;
-    grid-template-columns: 1fr 1fr 1fr 1fr ;
-    grid-gap: 5vw;
-}
