--- client/src/SideMenu.css
@@ -1,34 +1,0 @@
-* {box-sizing: border-box; margin: 0;}
-.widget {
-    padding: 20px 30px;
-    background: white;
-    font-family: 'Roboto', sans-serif;
-}
-.widget-title {
-    text-transform: uppercase;
-    letter-spacing: 2px;
-    color: #222;
-    font-size: 16px;
-    padding-left: 15px;
-    margin-bottom: 15px;
-    border-left: 2px solid #b99d61;
-}
-.widget-list {
-    padding: 0;
-    list-style: none;
-}
-.widget-list a:before {
-    content: "\2014";
-    margin-right: 14px;
-}
-.widget-list a {
-    text-decoration: none;
-    outline: none;
-    display: block;
-    padding: 6px 0;
-    letter-spacing: 1px;
-    font-weight: 300;
-    color: #444;
-    transition: .3s linear;
-}
-.widget-list a:hover {color: #b99d61;}
