+++ client/src/components/Content/Content.css
@@ -1,0 +1,11 @@
+.Content {
+  display: flex;
+  min-height: 100vh;
+  justify-content: center;
+}
+.Content-product {
+  display: grid;
+  grid-template-rows: 1fr 1fr 1fr 1fr;
+  grid-template-columns: 1fr 1fr 1fr 1fr;
+  grid-gap: 5vw;
+}
