+++ client/src/components/Header/Header.css
@@ -1,0 +1,49 @@
+.Header-logo {
+  font-size: 30pt;
+}
+.Header-navigator {
+  border-bottom: 1px solid black;
+}
+.Header-navigator ul {
+  display: block;
+  text-align: center;
+  margin: 0px;
+}
+.Header-navigator li {
+  display: inline-block;
+  margin: 10px;
+  position: relative;
+}
+.number {
+  margin-left: 20px;
+}
+.Header-button {
+  border: 2px solid black;
+  height: 55px;
+  width: 55px;
+  background: none;
+  text-transform: uppercase;
+  font-weight: 500;
+  font-size: 14px;
+  font-family: open sans, sans-serif;
+  margin: 18px;
+  padding: 0;
+  text-align: center;
+}
+.Header-button:hover {
+  background: black;
+  color: white;
+  transition: all 0.5s;
+  cursor: pointer;
+}
+.Header-button:focus {
+  outline: none;
+}
+.Header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  height: 80px;
+  /*border-top: 1px solid black; */
+  border-bottom: 1px solid black;
+}
