+++ client/src/pages/App.css
@@ -17,9 +17,11 @@
 .App-intro {
   font-size: large;
 }
-.
-
-@keyframes App-logo-spin {
-  from { transform: rotate(0deg); }
-  to { transform: rotate(360deg); }
+. @keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
 }
